Fix #7 - Performance issue
On each client's fd we monitored input AND output all the time. When a client is
connected and do nothing it is always available for writing even if the server
have nothing to write. This impacted the call to epoll_wait which instantly
return and cause a CPU load of 100%.
The fix is to monitor fds in a select-like way : monitor on-demand.